+++ components/bal/voie-editor.js
@@ -17,7 +17,7 @@ import Form from '@/components/form'
 import FormInput from '@/components/form-input'
 import AssistedTextField from '@/components/assisted-text-field'
 import DrawEditor from '@/components/bal/draw-editor'
-import LanguesRegionalesForm from '../langues-regionales-form/langues-regionales-form'
+import LanguesRegionalesForm from '@/components/langues-regionales-form'
 
 function VoieEditor({initialValue, closeForm}) {
   const [isLoading, setIsLoading] = useState(false)
